Added upgrading and imported sign image

+++ Assets/Entity/Scripts/Hand/Definitions/AHandUpgradeDefinition.cs
@@ -8,8 +8,8 @@ namespace Entity.Scripts.Hand.Definitions
         [SerializeField] private int _Tier;
         public int Tier => _Tier;
 
-        [SerializeField] private float _Cost;
-        public float Cost => _Cost;
+        [SerializeField] private int _Cost;
+        public int Cost => _Cost;
         
         [SerializeField] private float _Modifier;
         public float Modifier => _Modifier;
